+++ src/buffer.c
@@ -5,7 +5,7 @@
 
 #include "core.h"
 
-int ab_append(struct ABuf *ab, const char *s, size_t slen) {
+uint8_t ab_append(struct ABuf *ab, const char *s, size_t slen) {
     char *new = realloc(ab->buf, ab->len + slen);
     if (!new) die("realloc failed, out of memory");
 
@@ -16,7 +16,7 @@ int ab_append(struct ABuf *ab, const char *s, size_t slen) {
     return EXIT_SUCCESS;
 }
 
-int ab_free(struct ABuf *ab) {
+uint8_t ab_free(struct ABuf *ab) {
     free(ab->buf);
     ab->buf = NULL;
     return EXIT_SUCCESS;
